Se termina configuracion para error cors, 401 y se usa srpingsecurity con encripacion de contraseñas

+++ src/main/java/com/sicap/sciap_seguimiento_backend/entity/Proyecto.java
@@ -40,7 +40,7 @@ public class Proyecto {
     private String po;
 
     @Column(name = "usuario_creador")
-    private Integer usuarioCreador;
+    private Long usuarioCreador;
 
     @Column(name = "orden")
     private Integer orden;
